Bug 1004871 part 6 - Make GetPositionInIteration return a ComputedTiming object; r=dholbert
This patch makes ElementAnimations::GetPositionInIteration return
a ComputedTiming object instead of just a time portion (time fraction).

Since the ComputedTiming object includes phase information, we can fix those
parts of EnsureStyleRule and GetEventsAt that were temporarily using the time
portion to guess if the animation might have finished or not.

UltraBlame original commit: 81079a94b199876dd7af18ae7077834713beef82

+++ layout/style/AnimationCommon.h
@@ -323,12 +323,12 @@ struct ElementAnimation
   bool mIsRunningOnCompositor;
 
   enum {
-    LAST_NOTIFICATION_NONE = uint32_t(-1),
-    LAST_NOTIFICATION_END = uint32_t(-2)
+    LAST_NOTIFICATION_NONE = uint64_t(-1),
+    LAST_NOTIFICATION_END = uint64_t(-2)
   };
   
   
-  uint32_t mLastNotification;
+  uint64_t mLastNotification;
 
   InfallibleTArray<AnimationProperty> mProperties;
